use double for generator with R uniforms

+++ inst/include/ZigguratR.h
@@ -80,12 +80,12 @@ namespace R {
             return RNOR;
         }
     private:
-        float fn[128];
+        double fn[128];
         int32_t hz;
         uint32_t iz;
         uint32_t jz;
         uint32_t kn[128];
-        float wn[128];
+        double wn[128];
 
         void init() {               // called from ctor, could be in ctor
             double dn = 3.442619855899;
@@ -100,30 +100,30 @@ namespace R {
             kn[0] = (uint32_t) ((dn / q) * m1);
             kn[1] = 0;
 
-            wn[0]   = (float) (q / m1);
-            wn[127] = (float) (dn / m1);
+            wn[0]   = (double) (q / m1);
+            wn[127] = (double) (dn / m1);
             
             fn[0] = 1.0;
-            fn[127] = (float) (exp(- 0.5 * dn * dn));
+            fn[127] = (double) (exp(- 0.5 * dn * dn));
         
             for (i = 126; 1 <= i; i--) {
                 dn = sqrt(- 2.0 * log (vn / dn + exp (- 0.5 * dn * dn)));
                 kn[i+1] = (uint32_t ) ((dn / tn ) * m1);
                 tn = dn;
-                fn[i] = (float) (exp( - 0.5 * dn * dn));
-                wn[i] = (float) (dn / m1);
+                fn[i] = (double) (exp( - 0.5 * dn * dn));
+                wn[i] = (double) (dn / m1);
             }
             return;
         }
 
         //inline float nfix(void) { return wn[34]; }
-        inline float nfix(void) {
-            const float r = 3.442620;
-            float x, y;
+        inline double nfix(void) {
+            const double r = 3.442620;
+            double x, y;
             
             for (;;) {
                 //  IZ = 0 handles the base strip.
-                x = (float) (hz * wn[iz]);
+                x = (double) (hz * wn[iz]);
                 if ( iz == 0 ) { 
                     do {
                         x = - log (UNI) * 0.2904764; 
@@ -141,7 +141,7 @@ namespace R {
                 hz = KISS;
                 iz = (hz & 127);
                 if (fabs(hz) < kn[iz]) {
-                    return ((float) (hz * wn[iz]));
+                    return ((double) (hz * wn[iz]));
                 }
             }
         }
